Change MessageBody to Message
Kinda working except initializing header.message_ptr isn't correct,
probably because after creating a Message on the stack it gets copied
and is moved, I think.

In messages that are created in the heap this probably isn't a problem,
UNTESTED, but it would be nice if messages could be created in the stack.