Add code which allows use of rdtsc.
This is "working" but when compiling in builtin.Mode.Debug the cpu_id
returned by rdtscp has non-identical cpu_id's most of the time. So added
a workaround to CycleCounter.read which would ignore cpu_id's in
builtin.Mode.Debug. But then ran into a compiler error that there was
no error path. So worked around that by assume there would "never" have
cpu_id's equal, which is probably false, but good enough for now.